+++ Assets/Scripts/Health.cs
@@ -4,15 +4,15 @@ using UnityEngine;
 
 public class Health : MonoBehaviour
 {
-    public int startingHealth = 100;
-    private int health = 100;
+    public float startingHealth = 100;
+    private float health = 100;
 
     private void Start()
     {
         health = startingHealth;
     }
 
-    public void TakeDamage(int damage)
+    public void TakeDamage(float damage)
     {
         //Debug.Log("Enemy taken damage: " + damage);
         health -= damage;
